remove appbar from signin and signup pages - #213
