+++ frontend/src/index.jsx
@@ -12,10 +12,11 @@ import SignIn from './views/SignIn';
 import SignUp from './views/SignUp';
 import Room, { roomLoader } from './views/Room';
 import Rooms from './views/Rooms';
-import ErrorPage from './views/errorPage';
+import ErrorPage from './views/ErrorPage';
 import Home from './views/Home';
 import RoomNotFound from './views/RoomNotFound';
 import EditRoom from './views/EditRoom';
+import { RoomsLayout } from './layout/RoomsLayout';
 
 // import RoomTest from './components/RoomTest';
 import RequireAuth from './components/RequireAuth';
@@ -32,26 +33,42 @@ const router = createBrowserRouter([
         element: <Home />,
       },
       {
-        path: '/rooms/404',
-        element: <RoomNotFound />,
-      },
-      {
-        path: '/rooms/:roomId',
+        path: '/rooms',
         element: (
           <RequireAuth>
-            <Room />
+            <RoomsLayout />
           </RequireAuth>
         ),
-        loader: roomLoader,
-      },
-      {
-        path: '/rooms/:roomId/edit',
-        element: (
-          <IsAdmin>
-            <EditRoom />
-          </IsAdmin>
-        ),
-        loader: roomLoader,
+        children: [
+          {
+            path: '',
+            element: (
+              <RequireAuth>
+                <Rooms />
+              </RequireAuth>
+            ),
+          },
+          {
+            path: '404',
+            element: <RoomNotFound />,
+          },
+          {
+            path: ':roomId',
+            element: (
+              <Room />
+            ),
+            loader: roomLoader,
+          },
+          {
+            path: ':roomId/edit',
+            element: (
+              <IsAdmin>
+                <EditRoom />
+              </IsAdmin>
+            ),
+            loader: roomLoader,
+          },
+        ]
       },
       {
         path: '/signIn',
@@ -67,14 +84,6 @@ const router = createBrowserRouter([
           <RedirectLoggedInUser>
             <SignUp />
           </RedirectLoggedInUser>
-        ),
-      },
-      {
-        path: '/rooms',
-        element: (
-          <RequireAuth>
-            <Rooms />
-          </RequireAuth>
         ),
       },
       // {
