+++ frontend/src/views/ErrorPage.jsx
@@ -2,13 +2,18 @@ import React from 'react';
 import { useRouteError, useNavigate } from 'react-router-dom';
 import Button from '@mui/material/Button';
 
-export default function ErrorPage() {
+function ErrorPage() {
   const error = useRouteError();
   const navigate = useNavigate();
 
   return (
     <div style={{
-      width: '100vw', height: '100vh', display: 'flex', flexDirection: 'column', alignItems: 'center', justifyContent: 'center',
+      width: '100vw',
+      height: '100vh',
+      display: 'flex',
+      flexDirection: 'column',
+      alignItems: 'center',
+      justifyContent: 'center',
     }}
     >
       <h1>Oops!</h1>
@@ -20,3 +25,5 @@ export default function ErrorPage() {
     </div>
   );
 }
+
+export default ErrorPage;
